+++ backend/src/service/description_generator/json.rs
@@ -1,0 +1,12 @@
+use tracing::info;
+
+use super::{DescriptionGenerator, TemplateContext};
+use crate::Result;
+pub struct JsonDescriptionGenerator;
+
+impl DescriptionGenerator for JsonDescriptionGenerator {
+    fn generate(&self, options: TemplateContext) -> Result<String> {
+        info!("Generating JSON description");
+        serde_json::to_string_pretty(&options).map_err(From::from)
+    }
+}