+++ frontend/src/components/Loader.tsx
@@ -1,7 +1,8 @@
 import clsx from "clsx"
+import * as React from "react"
 
 interface Props {
-  children: React.ReactNode
+  children?: React.ReactNode
   className?: string
 }
 
@@ -9,7 +10,7 @@ const Loader: React.FC<Props> = ({children, className}) => {
   return (
     <div className={clsx("flex gap-4 items-center", className)}>
       <span className="loading loading-ring w-16" />
-      <p>{children}</p>
+      {children && React.Children.count(children) >= 1 && <p>{children}</p>}
     </div>
   )
 }